fix: Override handlers in new app communicator (#3570)
